+++ Citas/ServiceCategory.cs
@@ -21,7 +21,7 @@ namespace Citas
             this.SessionTimes = new HashSet<SessionTime>();
         }
     
-        public short ServiceCategoryId { get; set; }
+        public long ServiceCategoryId { get; set; }
         public string Name { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
